Recog_Face/requirements.txt
# ============================================
# Face Recognition System Requirements (Windows, Python 3.10)
# ============================================

# ML Frameworks
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
scipy==1.11.3

# Computer Vision
opencv-python==4.8.0.74
Pillow==12.0.0

# Face Recognition
deepface==0.0.96
keras-vggface==0.6

# Object Detection (YOLO) - Optional, not required for doorbell recognition
# Commented out due to torch/tensorflow typing-extensions conflict
# ultralytics==8.3.235
# torch==2.4.1
# torchvision==0.19.1

# API Server
Flask==3.1.2
flask-cors==6.0.1

# Firebase Integration
firebase-admin==6.0.0
google-cloud-storage==2.10.0

# Supabase + visitor embedding fetch
supabase==2.6.0

# Face recognition distances/encodings for doorbell matching
face_recognition==1.3.0

# Utilities
requests==2.32.5
tqdm==4.67.1
python-dotenv==1.0.0

# ============================================
# Notes
# ============================================
# - Place Firebase service account at backend/config/firebase-service-account.json
# - CPU-only installs work; GPU optional for PyTorch/YOLO.
# - Tested on Python 3.10.x, Windows 10/11.
# - face_recognition depends on dlib; ensure build tools/AVX support on Windows.
